Minor refactoring to modernize some code
- small refactorings made while doing debugging elsewhere

- modernize some code by removing explicit types and using auto

- change batcher construction to used braced list initialization {…} rather than function call syntax (…)

- deal with problems caused by implicit narrowing casts now disallowed by use of {…}, mostly by converting types to size_t where this should have been used anyway

+++ CppTransport/translator/backends/infrastructure/file_handling/buffer.cpp
@@ -43,7 +43,7 @@
 #define BUFFER_MAGIC_TAG "MAGIC_TAG"
 
 
-buffer::buffer(boost::filesystem::path fn, unsigned int cp)
+buffer::buffer(boost::filesystem::path fn, size_t cp)
 	: in_memory(false),
 	  filename(std::move(fn)),
 		capacity(cp),
